third_party/WebKit/LayoutTests/inspector/sources/debugger/debugger-pause-on-blocked-script-url-expected.txt

CONSOLE ERROR: line 1: Refused to execute JavaScript URL because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-...'), or a nonce ('nonce-...') is required to enable inline execution.

Tests that pause on exception works for inline script URLs blocked by CSP.

Script execution paused.
Call stack:
    0) testAction (pause-on-blocked-script-url.js:6)
Script execution resumed.
